Feat/payment

+++ E-Commerce-Platform-Ass1.Service/Services/CheckoutService.cs
@@ -15,27 +15,35 @@ namespace E_Commerce_Platform_Ass1.Service.Services
         private readonly ICartItemRepository _cartItemRepository;
         private readonly IOrderRepository _orderRepository;
         private readonly IOrderItemtRepository _orderItemtRepository;
+        private readonly IPaymentRepository _paymentRepository;
 
-        public CheckoutService(
-            ICartRepository cartRepository,
-            ICartItemRepository cartItemRepository,
-            IOrderRepository orderRepository,
-            IOrderItemtRepository orderItemtRepository
-        )
+        public CheckoutService(ICartRepository cartRepository, ICartItemRepository cartItemRepository,
+            IOrderRepository orderRepository, IOrderItemtRepository orderItemtRepository,
+            IPaymentRepository paymentRepository)
         {
             _cartRepository = cartRepository;
             _cartItemRepository = cartItemRepository;
             _orderRepository = orderRepository;
             _orderItemtRepository = orderItemtRepository;
+            _paymentRepository = paymentRepository;
         }
 
-        public async Task<Order> CheckoutSuccessAsync(Guid userId, string shippingAddress)
+        public async Task<Order> CheckoutSuccessAsync(Guid userId, string shippingAddress, List<Guid> selectedCartItemIds)
         {
             var cart = await _cartRepository.GetCartByUserIdAsync(userId);
 
             if (cart == null || !cart.CartItems.Any())
             {
                 throw new Exception("Cart is empty");
+            }
+
+            var selectItems = cart.CartItems
+                .Where(ci => selectedCartItemIds.Contains(ci.Id))
+                .ToList();
+
+            if (!selectItems.Any())
+            {
+                throw new Exception("No selected items");
             }
 
             var order = new Order
@@ -45,28 +53,39 @@ namespace E_Commerce_Platform_Ass1.Service.Services
                 ShippingAddress = shippingAddress,
                 Status = "PAID",
                 CreatedAt = DateTime.Now,
-                TotalAmount = cart.CartItems.Sum(ci => ci.ProductVariant.Price * ci.Quantity),
+                TotalAmount = selectItems.Sum(ci =>
+                    ci.ProductVariant.Price * ci.Quantity)
             };
 
             await _orderRepository.AddAsync(order);
 
-            var orderItems = cart
-                .CartItems.Select(ci => new OrderItem
-                {
-                    Id = Guid.NewGuid(),
-                    OrderId = order.Id,
-                    ProductVariantId = ci.ProductVariantId,
-                    ProductName = ci.ProductVariant.VariantName,
-                    Price = ci.ProductVariant.Price,
-                    Quantity = ci.Quantity,
-                })
-                .ToList();
+            var orderItems = selectItems.Select(ci => new OrderItem
+            {
+                Id = Guid.NewGuid(),
+                OrderId = order.Id,
+                ProductVariantId = ci.ProductVariantId,
+                ProductName = ci.ProductVariant.VariantName,
+                Price = ci.ProductVariant.Price,
+                Quantity = ci.Quantity
+            }).ToList();
 
             await _orderItemtRepository.AddRangeAsync(orderItems);
 
-            await _cartItemRepository.DeleteByCartIdAsync(cart.Id);
+            var payment = new Payment(order.Id, order.TotalAmount);
 
-            await _cartRepository.DeleteAsync(cart);
+            await _paymentRepository.AddAsync(payment);
+
+            await _cartItemRepository.DeleteByIdsAsync(
+                selectItems.Select(ci => ci.Id).ToList());
+
+            var remainingItems = cart.CartItems
+                .Where(ci => !selectedCartItemIds.Contains(ci.Id))
+                .ToList();
+
+            if (!remainingItems.Any())
+            {
+                await _cartRepository.DeleteAsync(cart);
+            }
 
             return order;
         }
